code cleanup

+++ firmware.c
@@ -62,7 +62,7 @@ FAST unsigned char CheckFirmware(char *name)
     if (f_open(&file, name, FA_READ) == FR_OK)
     {
         Error = ERROR_INVALID_DATA;
-        iprintf("Upgrade file size     : %llu\r", f_size(&file));
+        iprintf("Upgrade file size     : %lu\r", (unsigned long)f_size(&file));
         iprintf("Upgrade header size   : %lu\r", (unsigned long)sizeof(UPGRADE));
 
         if (f_size(&file) >= sizeof(UPGRADE))
@@ -123,8 +123,8 @@ FAST unsigned char CheckFirmware(char *name)
                         else iprintf("ROM CRC mismatch! from header: %08lX, calculated: %08lX\r",
                             rom_crc, ~crc);
                     }
-                    else iprintf("ROM size mismatch! from header: %lu, from file: %llu\r",
-                        pUpgrade->rom.size, f_size(&file)-sizeof(UPGRADE));
+                    else iprintf("ROM size mismatch! from header: %lu, from file: %lu\r",
+                        pUpgrade->rom.size, (uint32_t)(f_size(&file)-sizeof(UPGRADE)));
                 }
                 else iprintf("Invalid upgrade file header!\r");
             }
@@ -133,7 +133,7 @@ FAST unsigned char CheckFirmware(char *name)
           }
           else iprintf("Error creating linkmap\r");
         }
-        else iprintf("Upgrade file size too small: %llu\r", f_size(&file));
+        else iprintf("Upgrade file size too small: %lu\r", (uint32_t) f_size(&file));
         f_close(&file);
     }
     else iprintf("Cannot open firmware file!\r");
